requirements.txt
peft==0.6.2
transformers==4.38.2
diffusers>=0.23.1
python-telegram-bot==21.0.1
numpy==1.24.2
open_clip_torch